+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## <next>
 * Added multi-select filter
+* Added pagination footer and changed requesting/sorting/filtering of paginated data
 
 ## 7.3.14
 * Fixed filtering data in paginated grid