+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## <next>
 * Added multi-select filter
 * Added pagination footer and changed requesting/sorting/filtering of paginated data
+* Fixed currency&float filtering
 
 ## 7.3.14
 * Fixed filtering data in paginated grid