+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Added multi-select filter
 * Added pagination footer and changed requesting/sorting/filtering of paginated data
 * Fixed currency&float filtering
+* Fix to keyboard navigation
 
 ## 7.3.14
 * Fixed filtering data in paginated grid